[FEAT] 공지사항 상세 조회 (#316)
* fix: 쿼리 수정

- 공동 등수도 하나로 처리

* feat: 친구 확인 validate 추가

* fix: 테스트 수정

- path variable 이름 수정

* feat: notice_detail 엔티티 추가

- publish_status추가
- 기존 notice API 쿼리 변경

* feat: 상세조회 API 추가

* fix: 타입 변경

- 테스트 추가

* fix: 버그 수정

- html 반환
- enum 오류
- PathVariable 수정

* test: 테스트 수정

+++ src/main/java/ku_rum/backend/domain/notice/dto/response/NoticeResponse.java
@@ -5,7 +5,7 @@ import ku_rum.backend.domain.notice.domain.Notice;
 
 public record NoticeResponse(
         Long id,
-        Integer categoryId,
+        Long categoryId,
         String categoryName,
         String title,
         String link,
